backend/requirements.txt
google-genai>=0.3.0
fastapi>=0.104.0
websockets>=12.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0
python-dotenv>=1.0.0
numpy>=1.24.0